+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Normalize countries to common short names instead of official names, drop unmatched countries, and de-duplicate list
+- Add UN regions based on countries
 
 ## [1.0.0] - 2024-12-09